requirements.txt
torch~=1.8.1
pytorch-ignite
sklearn~=0.0
arrow
tilse
sentence-transformers
numpy~=1.20.3
scikit-learn~=0.24.2
transformers~=4.6.1